requirements.txt
beautifulsoup4==4.9.3
faiss-cpu==1.6.3
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
numpy==1.20.2
pandas==1.1.5
Pillow==8.2.0
psycopg2-binary==2.7.7
regex==2021.4.4
requests==2.25.1
sentence-transformers==1.1.0
tokenizers==0.10.2
torch==1.8.1 -f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.9.1 -f https://download.pytorch.org/whl/torch_stable.html
transformers==4.5.1
urllib3==1.26.4
validators==0.18.2
python-dotenv